Home mis à jour

+++ app/Database/Migrations/2025-11-03-060058_Categories.php
@@ -10,7 +10,7 @@ class Categories extends Migration
     {
         $this->forge->addField([
             'id' => [
-                'type' => 'BIGINT',
+                'type' => 'INT',
                 'constraint' => 20,
                 'unsigned' => true,
                 'null' => false,
